requirements.txt
async-generator==1.10
attrs==23.2.0
autopep8==2.0.4
certifi==2023.11.17
charset-normalizer==3.3.2
clize==5.0.2
colorclass==2.2.2
ConfigUpdater==3.1.1
datatable @ git+https://github.com/h2oai/datatable@a68ff5c2a913e200811f2e55edf078e53ab945e0
docopt==0.6.2
docutils==0.20.1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
h11==0.14.0
idna==3.6
molgenis-py-client==2.4.1
numpy==1.26.3
od==2.0.2
openpyxl==3.1.2
outcome==1.3.0
packaging==23.0
pandas==2.2.0
pip-upgrader==1.4.15
platformdirs==2.6.2
pycodestyle==2.11.1
PyScaffold==4.4
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2022.7.1
PyYAML==6.0.1
requests==2.28.2
selenium==4.17.2
setuptools-scm==7.1.0
sigtools==4.0.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
terminaltables==3.1.10
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.66.1
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
wsproto==1.2.0
XlsxWriter==3.1.9
yamlemxconvert==1.3.0
